refactor: PositiveNumber 객체 구현
